Initial commit of compound nodes.
This uses a new approach where the backend has an explicit, but
optional sub-component name for each node. This allows the frontend
to create nodes comprising of several backend (basi) nodes, but
group them with the same main node name.

The frontend has a `NodeRef` type that will allow parameters
and recorders to easily reference a simple node by a string or
a specific node by a node and sub-component name.