feat: Add new SIMD IPM solver.
New solver using SIMD types from the portable_simd project. This
currently requires nightly to build. It is a port of the existing
OpenCL IPM solver to Rust using SIMD. It is written using f64x4
types which will require the AVX2 instruction set.

Included is a genral refactor of the OCL IPM solver to improve
feature gating and code sharing with the SIMD solver.

Some tweaks to benchmark code.