chore: Update to Clp master and add Windows to CI. (#57)
This fixes a build issue on Windows. Refactor the
clp-sys build.rs to be more generic and add
MSVC specific build flags.

Use static version of HDF5 for easier install on
Windows, but this requires cmake to be installed.

Fix path escaping which was preventing the test from
working correctly on Windows.