feat: Add support for multi-model simulation.
Add `PywrMultiModel` schema and corresponding `MultiNetworkModel` to pywr-core. This update refactors the schema of the regular model to use a new "network" key. The schema then provides a `PywrNetwork` containing only the nodes, edges, parameters, metric sets and outputs. This is then reused by the regular model and the multi-model.

A new parameter is added that can receive a parameter from another model. This is implemented via the transfers in the multi-model configuration. This is fairly generic because it uses the metric system to allowing passing of node or parameter data from one model to another.

TODO:

- [ ] Write some documentation on the new structs.
- [ ] Check naming convention for inter-model transfers.
- [ ] Add a multi-model test that uses circular inter-model transfers.