feat: Initial commit of DerivedMetric
A derived metric is updated after solve and is for values
that could be dependent on state AND paramter values. These need
updating and storing for use between time-steps.

Adds a test for piecewise storage model that shows how the volume
is updated at the end of time-step, and the derived proportional
volume is calculated and retained for use in the next time-step
despite the max volumes being updated.

Fixes #63.