feat: Add RbfProfileParameter (#55)
An initial implementation of RbfProfile parameter. It uses a basic RBF interpolation procedure.

It also implements `VariableParameter` trait. This has required refactoring the `VariableParameter` trait into a generic and then providing an implementation for f64 and u32. This mimics the float and integer APIs from Pywr v1.x.

`TryFromV1Parameter` is implemented for RbfProfileParameter following an update to v1 schema v0.9.0.