feat: Improve design of inter-network transfers.
Remove the need for placeholder parameters to receive inter-model
transfer values. Instead a metric is used to indicate the name
of the inter-model transfer value to use at that particular
location. This simplifies the configuration of inter-model
transfers, but requires the build methods in the schema to recieve
a reference to any transfers.