feat: Initial implementation of NodeAttribute
This is a refactor of NodeReference to specify a schema facing
attribute (NodeAttribute). This idea is to remove the need for
the user of the schema to understand the internals of complex
nodes. Instead they will specify a node by its name in the schema
and an associated NodeAttribute.

NodeAttribute is intended as a global enum of all possible values.
However, each node may only support a subset of these attributes.
I.e. some of the attributes do not make sense for certain node
types.

This has largely worked. However, there is an ugly requirement
in the control curves to specify the NodeReference struct
where it was previously just a string.