feat: Improve model building from schema.
Combine the loading of parameters and nodes into a single
while loop. The intention is that each are attempted to be
loaded, falling back if a node or parameter they depend on
can not be found. They are then tried again once the load
has progressed.

The change allows for the removal of the `set_constraints`
method on `Node`. The functionality in this method is
added into the `add_to_model` method instead. This means
developers need to be a little careful with writing that
method.

The main problem is when a storage node depends on a
parameter for maximum volume. This requires that parameter
to be available (and evaluated) prior to the node in order
for the initial volume to work correctly.

Some basic developer documentation on how to write new
nodes is added to help with the above.