chore(depds)!: Upgrade to PyO3 v0.22.
This is a breaking change that removes the derive Clone from
NetworkState. The derive is not possible with PyObjects and
therefore it is not possible to clone the internal state
of PyParameter.

Pywr itself does not need NetworkState to be cloneable, but it
might be useful for users. Some workarounds may exist to
either serialize the PyParameter state or copy it via Python.
If these are required in future then Clone will be reintroduced.