feat: Implement deny_unknown_fields for many types.
Enforce `deny_unknown_fields` in parameters and node schemas. This
will help finding schema errors and avoid silently ignoring incorrect
or misnamed fields. This has required not flattening the meta
objects in the JSON schema. Therefore, the schema has changed with
the meta data now nested in a "meta" field. The meta objects do
not implement `deny_unknown_fields` which allows this object
to contain open ended data that might be useful for other tools.

Fixes #165.