feat: Initial commit of a mutual exclusivity node. (#198)
This is the initial support for using binary variables to apply a mutual
exclusivity constraint. I.e. flow is allow through up to N nodes
at a time. Aggregated node is refactored to allow
support for sets of nodes. These are typically created when a
compound node (e.g. PiecewiseLink) is added to an AggregatedNode.
The compound node creates multiple "core" nodes, and these should
all be constrained by the AggregatedNode's factors or exclusivity.

The same consideration applies to the virtual storage nodes.

To do this there is a new method on Node in the schema to retrieve
the core node indices which that node has added to the core model.
For now this is a static implementation, but there are some nodes
which the user may to configure how this is done. E.g. should a
LossLink be constrained on its net or gross flow.

Renames aggregated node `Factors` to `Relationship`.

This starts to resolve issue #187.