feat: Remove inline parameters. (#280)
This is the first step in resolving #249. Remove inline-parameter
support in the v2 schema. This requires some more complicated
logic for conversion from the v1 schema. The conversion method
is refactored and updated to accommodate moving v1 inline
parameters to the global parameter list in v2. The same method
is also used for conversion to timeseries.