feat: Add routing to RiverNode.
RiverNode now supports Delay and Muskingum based routing. This
also adds Muskingum parameters that calculate the factors in
the Muskingum equation and uses this in an aggregated node
relationship.

This requires updating the relationship constraint to support
a non-zero right-hand side.

Added basic information the book about how this works.