feat: Use u64 instead of usize for index parameters. (#323)
`u64` is more appropriate for data that should be the same across
multiple platforms. `usize` is platform specific.

This does not change the use of `NonZeroUsize` for the rolling
window size in virtual storage nodes. Ultimately this values is
used to size an array and that maximum size will be platform
dependent.

Currently there is some casting to `usize`. A separate update
could warn on the `cast_lossless` lint and remove the use of
`as` throughout.

One outstanding query is whether this should actually be i64
to support negative values?

Issue #291.