feat: A scenario sclicing and combinations.
This adds support for running subsets of the global scenario
domain. It implements slicing or user defined combinations. This
comes with an update to the JSON schema to put both "groups" and
"combinations" keys in the "scenarios" object. This is a revision
from v1 where "scenario_combinations" was a top-level key in the
JSON document.

In core the scenario index now stores the simulation based
indices and original schema (global) indices and labels. This
is useful for providing the correct global context in output
files.

Those parameters which use scenario indexing for looking up
data now do so using the simulation (i.e. subset) indices, and
therefore assume the data (e.g. a 2D array) is correctly sliced
for the given domain. This has been done for the parameters
that exist, but may need some additional work to be helpful
for custom parameters.

Updated CSV and HDF output formats now include concatenated
label from the global context.

New tests for slicing and combinations.