feat: Use wide crate for stable SIMD support. (#376)
This removes the need for the nightly Rust compiler to use the
IPM SIMD solver. This is achieved, with safe code, by using the
`wide` crate. However, it does mean that the solver is now not
generic with regard the SIMD type and number of lanes. Instead
it is now hard-coded to be f64x4.

Ensure ipm-simd and ipm-ocl are included in CI builds and tests.
Remove the nightly builds as no longer needed separately. Fix
some associated linting issues.