feat: Add node components. (#431)
Add a new `NodeComponent` enum that is used to determine
which sub-nodes are used in flow constraints. Each schema node
specifies a sub-set enum, similar to the attributes, which is then
used to return different core node indices as required.

To allow this there are now `NodeAttrReference` and
`NodeComponentReference` which are used in different
places depending on the context. Typically the aggregated
and virtual nodes will take component references now.

There is additional module level documentation on attributes
and components, and improved docstrings for several schema
nodes.

The attribute and component enums are made public, which
may also be useful for tools to determine which variants
are allowed.

Added some additional test models for aggregated storage
and virtual storage. The aggregated storage tests ensure
that these work correctly with all the main storage nodes.

Added new invalid model tests framework. This is to test
models which have a valid schema, but should error when
trying to "build" into a core model. Currently there is
only one invalid model to ensure that a flow node can
not be used in an `AggregatedStorageNode`.

Fixes #362