feat: Add new model and network level error types.
Improve the handling of errors during model building. This
will provide more context information about which part and
component of the build process failed.

In addition, it should pass the "cause" of the resulting
exception as any underlying `PyErr` (e.g. from a custom
parameter, or loading via pandas) if one exists.