refactor: Refactor and merge aggregation functions.
There were two f64 aggregation function enums in pywr-core.
These are merged together and shared between parameters
and metric sets. The refactor moves the aggregation enums
to a common module in both core and schema.

This also moves to a tagged enum representation for the
variants. This will allow more complex functions, that require
user data, to be implemented later.