feat: Add seasonal virtual storage reset. (#474)
Adds an active flag to virtual storage nodes. When not
active the storage bounds become unbounded, the node is
not updated with flows from the model, and it is reset.

This means the node will be at reset volume outside the
active period.

The active period is configured when selecting a Seasonal
reset period in the schema.

Add two new test models.