feat: Add routing to RiverNode. (#471)
RiverNode now supports Delay and Muskingum based routing. This
also adds Muskingum parameters that calculate the factors in
the Muskingum equation and uses this in an aggregated node
relationship.

- This requires updating the relationship constraint to support
a non-zero right-hand side.
- Added a new relationship type for specifying an coefficients
directly in an equation type constraint.
- Made this available in AggregatedNode.
- Use it for the Muskingum routing in RiverNode, so that it is
much clearer how this relates to the equation.
- Refactored the aggregated node constraints in the LP builder
and added documentation to aid the understanding of the
implementation here.
- When applying aggregated node coefficients to the LP use the
factors as provided. This means for ratio and proportional
factors, the preceeding code must make the second factor
negative. This improves the logical and readibility of
how these factors are applied.

Added basic information the book about how this works.