feat: Add model timing information to result structs. (#510)
A little bit of refactoring the lifetime of the timing
data. It matches the state a bit now. The timing data
is attached to the model results, and is also made
available to the Python extension. There is only
basic duration and speed information exposed currently.

Add __repr__ for model results.