fix: Swap "rows" and "cols" keys in table lookup definition.
When using row orientated data you should specify the number
of columns to use as the index. And vice-versa for column
orientated data.

The change also impacts the mean of rows and cols keys in the
"Both" lookup type. The internal function has been modified to
follow this updated naming convention.

Fixes #556.