feat: Allow parameters to (optionally) return value in after.
This refactors the calculation of parameters to allow them to
optionally return values from their "after" method. This can
be used to return post-solve values (e.g. deficits). It allows
parameters to return one value in "before" (e.g. demand required)
and another outcome in "after" (e.g. yield or power). This allows
the same data to be reused in the parameter struct (or object).

To do the above the "calc" method has been renamed to "before"
for consistency. Also derived metrics are now redundant and have
been removed.

The internal state for managing storage has also be refactored
to store the max_volume alongside the current volume. This makes
it easier to ensure the proportional volume is computed correctly
and only updated after the solve (or if volume is set directly).
I.e. it makes sure that the proportional volume does not change
during "before" if max volume is a parameter value.

A new listing showing how to use this feature in a Python
parameter is given.